feat: implement 22-role audit findings (free tier)
Backend (apps/api):
- Add pino structured logger replacing all console.* calls
- Add helmet middleware (security headers on every response)
- Add express-rate-limit (200 req/15min global)
- Add HMAC-SHA256 auth for device-sync and device-stats endpoints
  (advisory: warns + allows if DEVICE_HMAC_SECRET not configured)
- Change POST /loop from insert to upsert on (session_id, error_hash)
- Add POST /metrics/event funnel tracking route (5 event types)
- Fix GET /summary topErrorHashes to fetch only error_hash column
- Fix health check version to read npm_package_version env var
- Add .env.example documenting all required environment variables

Database (supabase/migrations):
- 20260629000000: Fix time_wasted_ms to BIGINT in sessions + loops
- 20260629000001: Add missing created_at column to sessions
- 20260629000002: Add funnel_events table with CHECK constraint + indexes
- 20260629000003: Add UNIQUE(session_id, error_hash) on loops for upsert
- Fix schema.sql reference: time_wasted_ms now BIGINT in both tables

Web (apps/web):
- Privacy policy page with full GDPR coverage, LandingFooter, correct date

Tests (packages/core):
- detectLoop.extra.test.ts: 7 additional LoopDetector tests covering
  deterministic time-window boundary, lastSeen advancement, re-detection
  after resolve, getAllLoops with resolved status

Tooling:
- eslint.config.js: ESLint v9 flat config with no-console: error
- scripts/set-version.sh: atomic version bump across all package.json + Cargo.toml
- CHANGELOG.md: Keep-a-Changelog format covering v1 through 2.8.2

+++ apps/api/supabase/schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS public.sessions (
   started_at        TIMESTAMPTZ NOT NULL,
   ended_at          TIMESTAMPTZ,                        -- null while session is active
   loops_detected    INTEGER     NOT NULL DEFAULT 0,
-  time_wasted_ms    INTEGER     NOT NULL DEFAULT 0,
+  time_wasted_ms    BIGINT      NOT NULL DEFAULT 0,
   tokens_saved      INTEGER     NOT NULL DEFAULT 0,
   file_types        TEXT[]      NOT NULL DEFAULT '{}',  -- ['ts', 'py'] — no paths
   extension_version TEXT        NOT NULL DEFAULT '0.1.0',
@@ -29,7 +29,7 @@ CREATE TABLE IF NOT EXISTS public.loops (
   session_id    TEXT        NOT NULL,                  -- references sessions.session_id
   error_hash    TEXT        NOT NULL,                  -- djb2 hash — no recoverable content
   occurrences   INTEGER     NOT NULL DEFAULT 1,
-  time_wasted_ms INTEGER    NOT NULL DEFAULT 0,
+  time_wasted_ms BIGINT     NOT NULL DEFAULT 0,
   file_type     TEXT        NOT NULL DEFAULT '',       -- 'ts', 'py' etc — no path
   status        TEXT        NOT NULL DEFAULT 'active'
                             CHECK (status IN ('active', 'resolved', 'ignored')),
